+++ src/co/edu/uniquindio/Nodo.java
@@ -2,6 +2,8 @@ package co.edu.uniquindio;
 
 import java.util.List;
 
+import co.edu.uniquindio.categoriassintacticas.CategoriaSintactica;
+
 public interface Nodo {
 
 	public Nodo getPadre();
@@ -11,5 +13,13 @@ public interface Nodo {
 	public List<Nodo> getHijos();
 	
 	public void setHijos(List<Nodo> hijos);
+	
+	public void recorrerArbol(List<RunnableNodo> runnables);
+	
+	public Nodo buscarAncestro(Class<? extends CategoriaSintactica> categoriaSintactica);
+	
+	public static interface RunnableNodo {
+		void run(Nodo nodo);
+	}
 
 }