+++ src/co/edu/uniquindio/Token.java
@@ -6,38 +6,62 @@ public class Token extends NodoImplemetado {
 	
 	private final Class<? extends CategoriaLexica> tipoToken;
 	private final String lexema;
-	private String valor;
+	private int linea;
+	private int columna;
+	private Simbolo simbolo;
 
 	public Token(Class<? extends CategoriaLexica> tipoToken, String lexema) {
 		this.tipoToken = tipoToken;
 		this.lexema = lexema;
 	}
 	
-	public Token(Class<? extends CategoriaLexica> tipoToken, String lexema, String valor) {
+	public Token(Class<? extends CategoriaLexica> tipoToken, String lexema, Simbolo simbolo) {
 		this.tipoToken = tipoToken;
 		this.lexema = lexema;
-		this.valor = valor;
+		this.simbolo = simbolo;
 	}
 	
 	public Class<? extends CategoriaLexica> getTipoToken() {
 		return tipoToken;
 	}
 	
-	public String getValor() {
-		return valor;
+	public Simbolo getSimbolo() {
+		return simbolo;
 	}
 	
 	public String getLexema() {
 		return lexema;
 	}
 	
-	public void setValor(String valor) {
-		this.valor = valor;
+	public void setSimbolo(Simbolo simbolo) {
+		this.simbolo = simbolo;
 	}
 	
 	@Override
 	public String toString() {
-		return String.format("[TipoToken: %s, Lexema: %s, Valor: %s]", tipoToken.getSimpleName(), lexema, valor);
+		return
+			String.format("[TipoToken: %s, Lexema: %s, Símbolo: %s, Línea: %s, Columna: %s]"
+			, tipoToken.getSimpleName()
+			, lexema
+			, simbolo
+			, linea
+			, columna);
+	}
+
+	public int getColumna() {
+		return columna;
+	}
+
+	public void setColumna(int columna) {
+		this.columna = columna;
+	}
+
+	public int getLinea() {
+		return linea;
+	}
+
+	public void setLinea(int linea) {
+		this.linea = linea;
 	}
 
 }