+++ src/co/edu/uniquindio/categoriaslexicas/CategoriaLexicaPalabras.java
@@ -27,7 +27,7 @@ public class CategoriaLexicaPalabras extends CategoriaLexicaBase {
 			}
 			
 			if(completoPalabra) {
-				return new Token(getClass(), lexema, getClass().getSimpleName());
+				return new Token(getClass(), lexema);
 			}
 			else {
 				analizadorLexico.regresarBacktracking();