+++ src/co/edu/uniquindio/categoriaslexicas/CategoriaLexicaUnCaracter.java
@@ -2,6 +2,9 @@ package co.edu.uniquindio.categoriaslexicas;
 
 import co.edu.uniquindio.Token;
 
+// TODO Refactorizar ésta categoría base para que no exista
+// y las categoría léxicas de éste tipo se manejen como cadenas de un solo
+// caracter de la clase CategoriaLexicaPalabras
 public class CategoriaLexicaUnCaracter extends CategoriaLexicaBase {
 	
 	protected char caracter;
@@ -15,7 +18,7 @@ public class CategoriaLexicaUnCaracter extends CategoriaLexicaBase {
 		else {
 			// Se acepta el token y se retorna.
 			analizadorLexico.irSiguienteCaracter();
-			Token token = new Token(getClass(), "" + caracter, getClass().getSimpleName());
+			Token token = new Token(getClass(), "" + caracter);
 			return token;
 		}
 	}